view_xifti_surface: Add option to draw borders if qualitative (useful for parcels). Minor tweak to dlabel behavior
Borders will be drawn between different-valued regions, not different-colored regions. So if the input is a dlabel file, and the mesh is colored by the label color table, the borders may subdivide the colored regions of the mesh.

If a dlabel cifti is used for `view_xifti_surface` , override the label colors if `colors` is provided