Search for "wb_path" during `ciftiTools.setOption` call, and never afterward
We removed the `wb_cmd` from functions, so we don't need to keep looking for the Workbench. Now, During `ciftiTools.setOption` the input path is checked and extended from the folder to the executable if applicable. Warning messages are raised if the path seems invalid, but ciftiTools will try to use it anyway.